screencast_common: split of SHM code into shm_util.c
