+++ src/screencast/pipewire_screencast.c
@@ -13,6 +13,7 @@
 #include <libdrm/drm_fourcc.h>
 
 #include "screencast.h"
+#include "shm_util.h"
 #include "wlr_screencast.h"
 #include "xdpw.h"
 #include "logger.h"
